分离实现文件  增加拖拽添加功能
